+++ launcher/DocumentWindow.cpp
@@ -109,6 +109,8 @@ void DocumentWindow::showEvent(QShowEvent *show_event)
     application()->dispatchPyMethod(m_py_object, "aboutToShow", QVariantList());
 
     setFocus();
+
+    application()->closeSplashScreen();
 }
 
 void DocumentWindow::resizeEvent(QResizeEvent *event)